Fix build issue
`localegen`, which is a part of the `locale` package, isn't installed in the
Ubuntu 16.04 container used as the base for this image. Thus the first RUN
command fails, and the container image is not built.

We need to install the package first, and then execute the `localegen`
command and set the environment variables.